rename similar to https://github.com/privacy-scaling-explorations/halo2curves/pull/95
