+++ src/lib.rs
@@ -1,5 +1,5 @@
 mod arithmetic;
-mod bernsteinyang;
+mod ff_inverse;
 pub mod fft;
 pub mod hash_to_curve;
 pub mod msm;
